backend/requirements.txt
fastapi
uvicorn
openai
requests
beautifulsoup4
python-multipart
pydantic
python-dotenv
lxml
PyPDF2